feat: Implement Project Dialog and Edit Dialog components
- Added ProjectDetail component to display project information with expandable description.
- Created ProjectDialog component to manage project details and editing functionality.
- Introduced ProjectEditDialog for editing project details with form validation.
- Added StatusDialog component for creating and updating project statuses.
- Updated projects page to integrate new dialogs and manage state for project and status actions.
