+++ src/app/(pages)/dashboard/projects/page.tsx
@@ -4,7 +4,6 @@ import { useEffect, useMemo, useState } from "react";
 import Column from "./_components/Column";
 import { Project, Status } from "./type";
 import Link from "next/link";
-import { Button, Dialog, DialogPanel, DialogTitle, Field, Input } from '@headlessui/react'
 import {
   DndContext,
   DragEndEvent,
@@ -21,17 +20,15 @@ import Card from "./_components/Card";
 import { SubmitHandler, useForm } from "react-hook-form";
 import clsx from "clsx";
 import { createStatus, deleteStatus, getCardData, getStatusData, updateProjectStatus, updateStatus } from "./Api";
-import ProjectDialog from "./_components/ProjectDialog";
+import ProjectDialog from "./_components/ProjectDialog/ProjectDialog";
 import MemberDialog from "./_components/MemberDialog";
 import SearchBox from "./_components/SearchBox";
+import StatusDialog from "./_components/StatusDialog";
 
 
-type Inputs = {
-  status: string
-}
+
 
 export default function projectsPage() {
-
   const [projects, setProjects] = useState<Project[]>([]);
   const [columns, setColumns] = useState<Status[]>([]);
   const [activeProject, setActiveProject] = useState<Project | null>(null);
@@ -39,16 +36,20 @@ export default function projectsPage() {
   const [projectTypes , setProjectTypes] = useState<string[]>([])
 
   const [isMemberDialogOpen, setIsMemberDialogOpen] = useState(false);
-  const [isNewStatusDialogOpen, setIsNewStatusDialogOpen] = useState(false)
+  const [isStatusDialogOpen, setIsStatusDialogOpen] = useState(false)
   const [isProjectDialogOpen, setIsProjectDialogOpen] = useState(false)
+
+  const [action, setAction] = useState('')
   
   const statusIds = useMemo(() => {
     return columns.map(columns => columns._id)
   }, [columns])
 
-  function toggleNewStatusDialog() {
-    setIsNewStatusDialogOpen(!isNewStatusDialogOpen)
-    reset();
+  function toggleStatusDialog(action: string, status: Status|null) {
+    setActiveColumn(status);
+    setAction(action)
+    setIsStatusDialogOpen(!isStatusDialogOpen)
+    
   }
 
   function toggleProjectDialog(){
@@ -57,29 +58,8 @@ export default function projectsPage() {
 
   function toggleMemberDialog() {
     setIsMemberDialogOpen(prev=>!prev)
-}
-
-  const {
-    register,
-    handleSubmit,
-    reset,
-    formState: { errors },
-  } = useForm<Inputs>();
-
-  const submitNewStatus: SubmitHandler<Inputs> = async ({ status }) => {
-    const newStatus = await createStatus(status);
-    if(!newStatus) return;
-    const {data} = newStatus;
-
-    setColumns((column) => {
-      const newColumn = {
-        _id: data._id,
-        title: status,
-      }
-      return [...column, newColumn]
-    })
-    toggleNewStatusDialog();
   }
+
 
   const sensors = useSensors(
     useSensor(PointerSensor, {
@@ -244,53 +224,11 @@ export default function projectsPage() {
               >
                 + Project
               </Link>
-              <div onClick={toggleNewStatusDialog}
+              <div onClick={()=>toggleStatusDialog('New',null)}
                 className="inline-flex items-center gap-2 rounded-md bg-blue-400 py-1.5 px-3 text-sm/6 font-semibold text-white shadow-inner shadow-white/10 focus:outline-none hover:bg-blue-700 focus:outline-1 focus:outline-white open:bg-gray-700 hover:cursor-pointer"
               >
                 + Status
               </div>
-              <Dialog open={isNewStatusDialogOpen} as="div" className="relative z-10 focus:outline-none" onClose={toggleNewStatusDialog}>
-                <div className='fixed inset-0 z-0 w-screen h-screen backdrop-blur-xs' />
-                <div className="fixed inset-0 z-10 w-screen overflow-y-auto ">
-                  <div className="flex min-h-full items-center justify-center p-4">
-                    <DialogPanel
-                      transition
-                      className="w-full max-w-md shadow border-1 rounded-xl bg-white p-6 duration-300 ease-out data-[closed]:transform-[scale(95%)] data-[closed]:opacity-0"
-                    >
-                      <DialogTitle as="h3" className="text-base/7 font-medium text-black">
-                        Enter new status name.
-                      </DialogTitle>
-                      <form onSubmit={handleSubmit(submitNewStatus)}>
-                        <Field>
-                          <Input
-                            type="text"
-                            {...register("status", { required: true })}
-                            className={clsx(
-                              'mt-3 block w-full rounded-lg border-1 bg-white/5 py-1.5 px-3 text-sm/6 text-black',
-                              'focus:outline-none data-[focus]:outline-2 data-[focus]:-outline-offset-2 data-[focus]:outline-white/25'
-                            )}
-                          />
-                          {errors.status && <span className="text-red-400">This field is required</span>}
-                        </Field>
-                        <div className="mt-4 space-x-3">
-                          <Button
-                            className="inline-flex items-center gap-2 rounded-md bg-blue-500 py-1.5 px-3 text-sm/6 font-semibold text-white shadow-inner shadow-white/10 focus:outline-none data-[hover]:bg-blue-700 data-[focus]:outline-1 data-[focus]:outline-white data-[open]:bg-gray-700 hover:cursor-pointer"
-                            type="submit"
-                          >
-                            Submit
-                          </Button>
-                          <Button
-                            className="inline-flex items-center gap-2 rounded-md bg-red-500 py-1.5 px-3 text-sm/6 font-semibold text-white shadow-inner shadow-white/10 focus:outline-none data-[hover]:bg-red-600 data-[focus]:outline-1 data-[focus]:outline-white data-[open]:bg-gray-700 hover:cursor-pointer"
-                            onClick={toggleNewStatusDialog}
-                          >
-                            Cancle
-                          </Button>
-                        </div>
-                      </form>
-                    </DialogPanel>
-                  </div>
-                </div>
-              </Dialog>
             </div>
           </div>
 
@@ -313,6 +251,7 @@ export default function projectsPage() {
                   setActiveProject={setActiveProject}
                   toggleProjectDialog={toggleProjectDialog}
                   toggleMemberDialog={toggleMemberDialog}
+                  toggleStatusDialog={toggleStatusDialog}
                 />
               ))}
             </SortableContext>
@@ -343,6 +282,7 @@ export default function projectsPage() {
               setActiveProject={()=> ({})}
               toggleProjectDialog={()=>({})}
               toggleMemberDialog={()=>({})}
+              toggleStatusDialog={()=>({})}
             />)
           }
         </DragOverlay>
@@ -361,6 +301,7 @@ export default function projectsPage() {
       )}
       
       <MemberDialog projectMember={activeProject?.users} isOpen={isMemberDialogOpen} toggleMemberDialog={toggleMemberDialog}/>
+      <StatusDialog isDialogOpen={isStatusDialogOpen} action={action} setColumns={setColumns} toggleStatusDialog={toggleStatusDialog} activeColumn={activeColumn??null} />
     </div>
   );
 }
